Additional pytests and restructuring of files in accordance with Al Madi's design
