+++ test/conftest.py
@@ -5,8 +5,8 @@ conftest.py
 import os
 import sys
 import pytest
-sys.path.append(os.path.abspath(
-    "/Users/jordansmith/Desktop/CS321/group-sprint-2/Mailroom-Flask-App"))
+# sys.path.append(os.path.abspath(
+#     "/Users/jordansmith/Desktop/CS321/group-sprint-2/Mailroom-Flask-App"))
 from website import create_app
 from website import models
